Mesh: - implemented support for UE3 SkeletalMesh with 32-bit index buffer - separated common parts of CSkelMeshSection/CStaticMeshSection to CMeshSection - made generic version of BuildTangents() - common for both mesh types - implemented export of meshes with more thank 64k vertices - separated common parts of SkeletalMesh and StaticMesh psk exporters - SkeletalMesh exporter will automatically use psk or pskx extension
Exporters:
- removed 2nd argument (FArchive&) from all exporter functions, removed file extension from
  exporter registrators
- added function CreateExportArchive() which is used from all exporter functions
- added option "-nooverwrite" to prevent re-exporting existing files
- tga export will create valid 1x1 tga file when UTexture::Decompress() is failed
- exporter will no more crash when unable to create file

Debugging:
- added compile option VSTUDIO_INTEGRATION which allows command-line switch "-debug"; when
  "-debug" is specified, umodel will try to launch debugger when crashed
- added ability to load symbols from pdb file; currently used for decoding EIP address only
- win32ExceptFilter() were moved to CoreWin32.cpp

Miscellaneous:
- added initial support for Batman: Arkham City (implemented package support)
- fixed FUE3ArchiveReader::GetFileSize() for Bioshock2 (fixed crash in decompressor tool)
- updated ActorXImporter.ms to version 1.19
- fixed compilation without UNREAL3 define
- FFileReader: added ability to not crash when file cannot be opened
- added DECLARE_STRUCT2 macro for structures with parents which has typeinfo


git-svn-id: svn://localhost@146 a3f5f06f-15b6-264c-8744-b656ee826ca1

+++ test.sh
@@ -1,5 +1,14 @@
 #!/bin/bash
 
+# Usage:
+#	test.sh [--nobuild] [--game=...] <umodel options>
+# Options:
+# --nobuild			use as 1st option to prevent from umodel rebuilding
+# --<game>			choose predefined game path
+# -path=<path>		will disable default game substitution
+#?? add --exe=<executable>, do not build when overrided (usefult for testing with older exe from svn)
+
+
 # Enable extended pattern matching (for (a|b|c)); see bash manpage, "Pattern matching".
 shopt -s extglob
 
@@ -75,8 +84,14 @@ function mass()   { run1 "$MASS" $*; }
 function dcu()    { run1 "$DCU" $*;  }
 function alice()  { run1 "$ALICE" $*; }
 
-rm $exe
-./build.sh
+# rebuild umodel when not desired opposite
+#!! WARNING: order-dependent option
+if [ "$1" == "--nobuild" ]; then
+	shift
+else
+	rm $exe
+	./build.sh
+fi
 
 # Check directories
 #?? should check dirs when specific game has been requested (not all games everytime)
